pyproject.toml
[tool.poetry]
name = "django-blickk-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.5"
django-extensions = "^3.1.0"
pytelegrambotapi = "^3.7.5"
python-dotenv = "^0.15.0"
flake8 = "^3.8.4"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
